backend/requirements.txt
fastapi
uvicorn
pandas
python-multipart
pytest
boto3
sqlalchemy
pymongo
dotenv
asyncpg
psycopg2
google-genai